Change the MonadIO instance again.
Wrapping every exception thrown from IO code into UncheckedIOException
has the inconvenience that now every handler must unwrap the exception.

This is confusing and surprinsing. You are no longer handling an
e.g. SQLException anymore, but an UncheckedIOException.

I'm sure with some extra magic in catch this could be worked around, but
for the moment I'm going to return to the old way of doing the MonadIO
instance, i.e. adding a Throws SomeException constraint, with a new twist:
now SomeException is an unchecked exception.